Separate lines from take_bounded_screenshot
Due to the lines being too many at function take_bounded_screenshot and
is seemingly unpractical in code readability, parts of this
function is split into its separate functions. This however only
separates lines that are instances of the class and can be safely
removed of its scope, further commits will change the local variables of
take_bounded_screenshot as well to further reduce the amount of lines.